added a pesky cast

+++ src/Ch03/03_06b/CodeDemo.cpp
@@ -13,7 +13,7 @@ int main(){
     health_items -= 2; // Player uses some health items
 
     std::cout << "Final Ammo: " << ammo << std::endl;
-    std::cout << "Remaining Health Items: " << health_items << std::endl;
+    std::cout << "Remaining Health Items: " << (int) health_items << std::endl;
 
     std::cout << std::endl << std::endl;
     return 0;
